BACKEND/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
